Update shipping.py
I Googled a way to find the smallest variable, then used that to find the lowest cost option.  I turned that into a string, added an empty string to create a blank line, then moved all the print statements to the bottom of the program.  This way the cheapest option is printed by itself, but the customer is still able to see what all of their options are.  And it only made the program one line longer than it already was.